Refactor: Refactor button files
