+++ core/ui/designsystem/src/main/java/com/example/core/ui/designsystem/components/button/PrivacyPolicyButton.kt
@@ -1,0 +1,71 @@
+package com.example.core.ui.designsystem.components.button
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.LocalRippleConfiguration
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.RippleConfiguration
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.tooling.preview.PreviewLightDark
+import androidx.compose.ui.unit.dp
+import com.example.core.ui.designsystem.R
+import com.example.core.ui.designsystem.theme.FitfitTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun PrivacyPolicyButton(
+    onClick: () -> Unit,
+){
+    val grayRippleConfiguration = RippleConfiguration(
+        color = MaterialTheme.colorScheme.onSurfaceVariant
+    )
+
+    CompositionLocalProvider(LocalRippleConfiguration provides grayRippleConfiguration) {
+        TextButton(
+            onClick = onClick
+        ) {
+            Text(
+                text = stringResource(id = R.string.privacy_policy),
+                style = MaterialTheme.typography.labelMedium.copy(
+                    color = MaterialTheme.colorScheme.onSurfaceVariant,
+                    textDecoration = TextDecoration.Underline
+                )
+            )
+        }
+    }
+}
+
+
+
+
+
+
+
+
+@Composable
+@PreviewLightDark
+private fun PrivacyPolicyPreview(){
+    FitfitTheme {
+        Box(
+            contentAlignment = Alignment.Center,
+            modifier = Modifier
+                .background(MaterialTheme.colorScheme.surface)
+                .padding(16.dp)
+                .width(190.dp)
+        ){
+            PrivacyPolicyButton(
+                onClick = {},
+            )
+        }
+    }
+}
