+++ core/ui/designsystem/src/main/res/values/string.xml
@@ -3,6 +3,11 @@
     <string name="example">example</string>
 
     <!--component/button/ =======================================================================-->
+    <!-- AccountButtons.kt -->
+    <string name="delete_account">Delete account</string>
+    <string name="change_image">Change image</string>
+    <string name="delete_image">Delete image</string>
+
     <!-- AuthButtons.kt -->
     <string name="sign_in_with_google">Sign in with Google</string>
     <string name="sign_in_with_apple">Sign in with Apple</string>
@@ -15,12 +20,18 @@
     <string name="save">Save</string>
     <string name="cancel">Cancel</string>
 
-    <!-- buttons.kt -->
-    <string name="privacy_policy">Privacy policy</string>
+    <!-- MainWorkoutButtons.kt -->
     <string name="start_workout">Start Workout</string>
+
+    <!-- PrivacyPolicyButton.kt -->
+    <string name="privacy_policy">Privacy policy</string>
+
+    <!-- TryAgainButton.kt -->
     <string name="try_again">Try again</string>
-    <string name="delete_account">Delete account</string>
-    <string name="change_image">Change image</string>
+
+    <!-- WorkoutButtons.kt -->
+    <string name="next_workout">Next workout</string>
+
 
     <!--component/topAppBars/ ===================================================================-->
     <!-- FitfitTopAppBar.kt -->
@@ -41,10 +52,12 @@
 
     <!--IconTextButtonIcon-->
 
+    <!--IconTextButtonIcon-->
+    <string name="stop">Stop</string>
+    <string name="pause">Pause</string>
 
     <!--MyIcons-->
     <string name="error">Error</string>
-    <string name="delete_image">Delete image</string>
 
 
 </resources>
